requirements.txt
tensorflow
keras
numpy
pandas 
matplotlib
ipykernel
pillow
scipy
PyQt5
scikit-learn
seaborn
opencv-python
